// backend/src/config/env.d.ts
export declare const env: {
    PORT: string;
    NODE_ENV: "development" | "production" | "test";
    SUPABASE_URL: string;
    SUPABASE_ANON_KEY: string;
    SUPABASE_SERVICE_ROLE_KEY: string;
    JWT_SECRET: string;
    JWT_EXPIRES_IN: string;
    CORS_ORIGIN: string;
};
//# sourceMappingURL=env.d.ts.map